+++ packages/react/package.json
@@ -71,7 +71,7 @@
   "dependencies": {
     "@accounts/client": "0.0.3",
     "@accounts/common": "0.0.3",
-    "immutable-form": "^0.0.12",
+    "immutable-form": "^0.0.14",
     "react": "^15.3.2",
     "react-redux": "^4.4.5",
     "react-router": "^3.0.0",
